LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-setRemoteDescription-rollback-expected.txt


PASS setRemoteDescription(rollback) in have-remote-offer state should revert to stable state
PASS setRemoteDescription(rollback) from stable state should reject with InvalidStateError
PASS setRemoteDescription(rollback) should ignore invalid sdp content and succeed
PASS local offer created before setRemoteDescription(remote offer) then rollback should still be usable
PASS local offer created before setRemoteDescription(remote offer) with different transceiver level assignments then rollback should still be usable
FAIL rollback of a remote offer should remove a transceiver assert_equals: expected 0 but got 1
FAIL rollback of a remote offer should remove touched transceiver assert_equals: expected 0 but got 1
PASS rollback of a remote offer should keep a transceiver
PASS rollback of a remote offer should keep a transceiver created by addtrack
PASS rollback of a remote offer should keep a transceiver without tracks
PASS explicit rollback of local offer should remove transceivers and transport
FAIL when using addTransceiver, implicit rollback of a local offer should visit stable state, but not fire negotiationneeded until we settle in stable assert_true: There should be no negotiationneeded event right now expected true got false
FAIL when using addTrack, implicit rollback of a local offer should visit stable state, but not fire negotiationneeded assert_true: There should be no negotiationneeded event in this test expected true got false
PASS rollback of a remote offer to negotiated stable state should enable applying of a local offer
PASS rollback of a local offer to negotiated stable state should enable applying of a remote offer
PASS rollback a local offer with audio direction change to negotiated stable state and then add video receiver
FAIL two transceivers with same mids promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to set remote offer sdp: Called in wrong state: have-local-offer"
PASS onremovetrack fires during remote rollback
FAIL rollback of a remote offer with stream changes assert_equals: expected 2 but got 1
PASS removeTrack() with a sender being rolled back does not crash or throw
FAIL Implicit rollback with only a datachannel works promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to set remote offer sdp: Called in wrong state: have-local-offer"
